refactor: 전체보기 페이지 컴포넌트를 DrinksLIstPage 컴포넌트로 변경, 검색 페이지에서 카테고리 선택 시 drinks로 요청 패스 변경
