+++ frontend/src/constants/path.ts
@@ -7,7 +7,6 @@ const PATH: { [key: string]: string } = {
   MY_DRINKS: '/mypage/drinks',
   MY_REVIEWS: '/mypage/reviews',
   DRINKS: '/drinks',
-  VIEW_ALL: '/drinks/all',
   SEARCH: '/search',
   SEARCH_RESULT: '/search/result',
   PREFERENCE: '/preference',
