+++ frontend/src/pages/DrinksListPage/index.tsx
@@ -7,15 +7,32 @@ import NavigationHeader from 'src/components/Header/NavigationHeader';
 import ListItem from 'src/components/Item/ListItem';
 import List from 'src/components/List/List';
 import { PATH } from 'src/constants';
-import { Container, InfinityScrollPoll } from './ViewAllPage.styles';
+import { Container, InfinityScrollPoll } from './styles';
 
-const ViewAllPage = () => {
+const CATEGORY_NAME = {
+  SOJU: '소주',
+  BEER: '맥주',
+  YANGJU: '양주',
+  MAKGEOLLI: '막걸리',
+  COCKTAIL: '칵테일',
+  WINE: '와인',
+  ETC: '기타 주류',
+  ALL: '전체보기',
+};
+
+const DrinksListPage = () => {
   const history = useHistory();
   const infinityPollRef = useRef<HTMLDivElement>(null);
 
+  const category =
+    (new URLSearchParams(history.location.search).get('category') as keyof typeof CATEGORY_NAME) ??
+    'ALL';
+  const categoryName = CATEGORY_NAME[category];
+  const params = new URLSearchParams({ category });
+
   const { data, fetchNextPage, hasNextPage, isFetching } = useInfiniteQuery(
     'drinks',
-    ({ pageParam = 1 }) => API.getDrinks({ page: pageParam }),
+    ({ pageParam = 1 }) => API.getDrinks({ page: pageParam, params }),
     {
       getNextPageParam: ({ pageInfo }) => {
         return pageInfo.currentPage < pageInfo.lastPage ? pageInfo.currentPage + 1 : undefined;
@@ -23,6 +40,7 @@ const ViewAllPage = () => {
     }
   );
   const drinks = data?.pages?.map((page) => page.data).flat() ?? [];
+  const totalSize = data?.pages[0].pageInfo?.totalSize;
 
   const onMoveToDrinkDetail = (id: number) => () => {
     history.push(`${PATH.DRINKS}/${id}`);
@@ -48,7 +66,7 @@ const ViewAllPage = () => {
 
   return (
     <Container>
-      <NavigationHeader title="전체보기" />
+      <NavigationHeader title={`${categoryName} (${totalSize ?? 0}개)`} />
       <List count={drinks?.length}>
         {drinks?.map((item: Drink.Item) => (
           <ListItem
@@ -70,4 +88,4 @@ const ViewAllPage = () => {
   );
 };
 
-export default ViewAllPage;
+export default DrinksListPage;
